Port AmigaOS4 OGLES2 support to axmol v2.11.3
Rebased all AmigaOS4+OGLES2 patches from v2.3.1 onto v2.11.3 (latest v2.x LTS).
450 upstream commits absorbed with 8 manual conflict resolutions.

Changes from v2.3.1 port:
- GLViewImpl → RenderViewImpl rename throughout
- messageBox → showAlert API (Common-amigaos4.cpp)
- axmolver.h generated for v2.11.3
- Opus source exclusion in ogg build
- 2-space indentation in CMake files

+++ core/platform/Image.cpp
@@ -2701,7 +2701,7 @@ bool Image::saveImageToJPG(std::string_view filePath)
         AX_BREAK_IF(nullptr == outfile);
 
         unsigned char* outputBuffer = nullptr;
-        unsigned long outputSize    = 0;
+        size_t outputSize           = 0;
         jpeg_mem_dest(&cinfo, &outputBuffer, &outputSize);
 
         cinfo.image_width      = _width; /* image width and height, in pixels */
